only need version check in tests
