+++ setup.cfg
@@ -31,8 +31,6 @@ install_requires =
     click>=7
     rich>=9
     dataclasses>=0.6;python_version<"3.7"
-    importlib-metadata;python_version<"3.8"
-    importlib-resources;python_version<"3.7"
 python_requires = >=3.6
 
 [options.entry_points]
@@ -43,6 +41,8 @@ console_scripts =
 tests =
     pytest>=6
     pytest-cov>=2.10
+    importlib-metadata;python_version<"3.8"
+    importlib-resources;python_version<"3.7"
 
 [mypy]
 files = dis_cli.py, tests/*.py
